requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
lxml>=5.0.0
readability-lxml>=0.8.1
httpx>=0.27.0
pyyaml>=6.0.1
jinja2>=3.1.4
tenacity>=8.3.0
